refactor(auth): add single-components gates for domain-workflows and clean types and naming (#1237)
reference: https://github.com/cadence-workflow/cadence-web/issues/1234

adresses:
- Move ERROR_SNACKBAR_OVERRIDES to a .styles.ts file
- Rename logoutReasonRef to "logout trigger" for clear intent
- Refactore the domain workflows page to follow single component per
file pattern

---------
