+++ src/components/app-nav-bar/app-nav-bar.styles.tsx
@@ -1,3 +1,5 @@
+import { type SnackbarElementOverrides } from 'baseui/snackbar';
+
 import type {
   StyletronCSSObject,
   StyletronCSSObjectOf,
@@ -12,3 +14,13 @@ const cssStylesObj = {
 
 export const cssStyles: StyletronCSSObjectOf<typeof cssStylesObj> =
   cssStylesObj;
+
+export const overrides = {
+  errorSnackbar: {
+    Root: {
+      style: {
+        backgroundColor: '#c62828',
+      },
+    },
+  } satisfies SnackbarElementOverrides,
+};
