+++ src/components/auth-token-modal/auth-token-modal.types.ts
@@ -1,0 +1,5 @@
+export type Props = {
+  isOpen: boolean;
+  onClose: () => void;
+  onSubmit: (token: string) => Promise<void> | void;
+};
